fixed UI bugs

+++ lib/screens/add_product_screen.dart
@@ -180,7 +180,7 @@ class _AddProductScreenState extends State<AddProductScreen> {
                         if (value!.isEmpty) {
                           return 'Please provide a value.';
                         }
-                        if(int.tryParse(value)!=null){
+                        if(double.tryParse(value)!=null){
                           return 'Product name can not be a number';
                         }
                         return null;
